+++ modules/apps/29-fee/fee_test.go
@@ -5,9 +5,16 @@ import (
 
 	"github.com/stretchr/testify/suite"
 
+	icacontroller "github.com/cosmos/ibc-go/v7/modules/apps/27-interchain-accounts/controller"
+	icacontrollertypes "github.com/cosmos/ibc-go/v7/modules/apps/27-interchain-accounts/controller/types"
+	icahost "github.com/cosmos/ibc-go/v7/modules/apps/27-interchain-accounts/host"
+	icahosttypes "github.com/cosmos/ibc-go/v7/modules/apps/27-interchain-accounts/host/types"
 	"github.com/cosmos/ibc-go/v7/modules/apps/29-fee/types"
+	"github.com/cosmos/ibc-go/v7/modules/apps/transfer"
+	transfertypes "github.com/cosmos/ibc-go/v7/modules/apps/transfer/types"
 	clienttypes "github.com/cosmos/ibc-go/v7/modules/core/02-client/types"
 	channeltypes "github.com/cosmos/ibc-go/v7/modules/core/04-channel/types"
+	porttypes "github.com/cosmos/ibc-go/v7/modules/core/05-port/types"
 	ibctesting "github.com/cosmos/ibc-go/v7/testing"
 	ibcmock "github.com/cosmos/ibc-go/v7/testing/mock"
 )
@@ -62,6 +69,40 @@ func (suite *FeeTestSuite) CreateMockPacket() channeltypes.Packet {
 		clienttypes.NewHeight(0, 100),
 		0,
 	)
+}
+
+// RemoveFeeMiddleware removes:
+// - Fee middleware from transfer module
+// - Fee middleware from icahost submodule
+// - Fee middleware from icacontroller submodule
+// - The entire mock-fee module
+//
+// It does this by overriding the IBC router with a new router.
+func RemoveFeeMiddleware(chain *ibctesting.TestChain) {
+	channelKeeper := chain.GetSimApp().IBCKeeper.ChannelKeeper
+
+	// Unseal the IBC router by force
+	chain.GetSimApp().IBCKeeper.Router = nil
+	chain.GetSimApp().IBCKeeper.PortKeeper.Router = nil
+
+	newRouter := porttypes.NewRouter() // Create a new router
+	// Remove Fee middleware from transfer module
+	chain.GetSimApp().TransferKeeper.WithICS4Wrapper(channelKeeper)
+	transferStack := transfer.NewIBCModule(chain.GetSimApp().TransferKeeper)
+	newRouter.AddRoute(transfertypes.ModuleName, transferStack)
+
+	// Remove Fee middleware from icahost submodule
+	chain.GetSimApp().ICAHostKeeper.WithICS4Wrapper(channelKeeper)
+	icaHostStack := icahost.NewIBCModule(chain.GetSimApp().ICAHostKeeper)
+	newRouter.AddRoute(icahosttypes.SubModuleName, icaHostStack)
+
+	// Remove Fee middleware from icacontroller submodule
+	chain.GetSimApp().ICAControllerKeeper.WithICS4Wrapper(channelKeeper)
+	icaControllerStack := icacontroller.NewIBCMiddleware(nil, chain.GetSimApp().ICAControllerKeeper)
+	newRouter.AddRoute(icacontrollertypes.SubModuleName, icaControllerStack)
+
+	// Override and seal the router
+	chain.GetSimApp().IBCKeeper.SetRouter(newRouter)
 }
 
 // helper function